pyproject.toml
[project]
name = "wukong-invite-helper"
version = "0.1.0"
description = "Monitor Wukong invite image updates, extract invite codes with OCR, and optionally autofill the Wukong app on macOS."
requires-python = ">=3.11"
dependencies = [
    "Pillow",
]

[project.optional-dependencies]
tesseract = ["pytesseract"]

[project.urls]
Homepage = "https://github.com/LeoLB-Wang/wukong-invite-helper"
Repository = "https://github.com/LeoLB-Wang/wukong-invite-helper"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
